:wrench: Convert Capstone project into an Xcode project
Although it would be a nice experiment to try building this project
using Swift Playgrounds 4/Xcode, it is currently limited to iOS.

After a careful assessment of the project objectives, I have decided
that I will instead perform this experiment more traditionally with
and Xcode project set up to be multiplatform, targeting both iOS
and macOS simiulatenously.

This also opens the door to using other Apple-provided technologies
such as Core Data, CloudKit, and Sign In with Apple. These toolkits
may not be used in the project necessarily, but Xcode will support
these features, unlike SP4 at the moment.

Signed-off-by: Marquis Kurt <software@marquiskurt.net>