:wrench: Add Starlight URI type to Xcode project
As of the 1.0 release of Chica, noted in the library documentation,
Chica only authenticates and operates on the starlight:// URI type.

This commit updates the Xcode configuration on the iOS target to
support this URI protocol, so that Chica can operate correctly and
fetch/send data against the Gopherdon instance.

Signed-off-by: Marquis Kurt <software@marquiskurt.net>