:construction: Add Chica OAuth flow from documentation
A simple login system has been constructed to test the authentication
system for Chica when the user opens the app. Currently, it merely
checks for authentication status and optionally fetches test data at the
end of the authentication loop. If the user has logged in before with the
app, the login process is skipped.

This requires that the macOS app have access to inbound and outbound
connections to make the appropriate calls, as well as fixing the identifier
for this project to what is registered in Apple Developer:
'net.marquiskurt.shout'.

A better design for the login page should be in the works.