:construction: Readjust pinstripes
Originally, the pinstripes did not look as intended from the wireframe.
Using the Path components inside of a GeometryReader proved to
be too difficult to create a consistent-looking set of pinstripes in the
graphic from the wireframe.

To resolve this, a new Pinstripes asset has been added to the app's
asset library, and through the use of some clever geometry, the image
is placed into the correct position and size on various device sizes.