:lipstick: Update authentication design
The design for the authentication UI has been tweaked a bit more to
fit the aesthetic of the platforms they represents. On macOS, the
authentication dialog opens in a sheet, and clicking "Get Started" will
open the browser, without any additional dialogs appearing. Also, the
background window displays an image of cliffs, specific to macOS to
fit the internal codename's theme.

On iOS and iPadOS, the designs remain as the typical ones from the
previous commit.