:white_check_mark: Add unit tests and String.toPlainText method
Unit tests have been added to the project to help ensure critical functions
work as expected. The first example test included is for the
String.toPlainText method included in the Utilities, introduced with this
commit. As the name suggests, this method attempts to convert a string
consisting of sanitized HTML code and converts it to plain text. Since the
app will be using plain text and other contextual information attached with
a status to get information, it is unnecessary to render the HTML.