:lipstick: Make general improvements to StatusView
StatusView is more configurable with different options for where to
place the creation date and the profile image. To adjust some of the
options, view modifiers for StatusView are available; this route was
chosen to declutter the view's init call.

Additionally, a new date formatter has been introduced to parse dates
written in the Mastodon format.