:sparkles: Load home timeline on widescreen layout
The widescreen layout supports fetching data from the home timeline,
which is ideal for iPadOS and macOS. This will be introduced into the
compact layout soon, once the view to make it possible has been
designed.

Currently, the app runs with about ~45MB in memory, which is important
to keep in mind when making future calls. Perhaps the number of posts
needs to be reduced down to ten (or maybe a user-defined setting)?