:recycle: Refactor parts of timeline interfaces
To make fetching mock data easier, a MockData struct has been
introduced, containing static fields that can be used to fetch the
appropriate mock data. Previews and dummy views have been
update to use this new struct.

A new StatusDetailView page has been added to show a more
detailed view of a status. Ideally, this page will also show replies
and other statuses related to the discussion, but the
implementation currently is not present.

Additionally, a new Localizable.stringsdict has been added to
include localized strings that follow plural rules. This is used in
StatusDetailView on macOS to show how many replies there are
for a status in the subtitle of the window.