:bug: Fix several bugs in timelines, authentication
In previous iterations of the prototype, it was impossible to sign
into Gopherdon on iPadOS due to a structural problem with the
authentication view. This has been fixed so that, when tapping
on "Get Started", the user is redirected to the login page, without
opening the dialog first.

A critical bug with converting an HTML string to plain text has been
fixed due to a threading issue with respect to NSAttributedString.

Finally, some smaller improvements to the status view have been
made, which affect the placement of items in the view. These small
changes have been made in an effort to make the view work in
different contexts across various platforms. Notably, the status view
in the StatusListMDView has been updated so that it appears more
like an email message in Mail.app, removing the avatar and moving
the creation date from the bottom right to the top right of the row.