:bug: Add external events to main Shout view
In previous commits and iterations, after Chica authenticates with a
Mastodon instance, it opens a new instance of the app receiving the
authentication tokens, causing a small mess of windows and state
management. To resolve this and force Chica to reuse windows, two
"handlesExternalEvents" modifiers have been added to listen for the
OAuth endpoint that Chica uses for authentication to hijack control of
the window, thus forcing it into a single window.