:sparkles: Add basic post construction and reply system
This commit introduces the utilities and views needed to allow users
to make posts and reply to other posts in a thread with relative ease.
On a Mac, when clicking on the "New Post" or "Reply" buttons, a new
window appears with the visibility tools and a character remaining view.
At the bottom, if the user is replying to a post, it will show the original
post in question for context, making it easier to reply without bouncing
between views. For iOS, the compose window will appear as an action
sheet instead.

The design doesn't match the wireframes exactly, and it is unknown
how to get the sheet to act more like a window in the same way email
message composition in Mail.app works, but it will suffice for now. In
addition, this view currently supports plain text and doesn't include those
mentioned in the thread like most Mastodon clients do. That could be
fixed in a future commit.