:bug: Fix state problems with respect to reply IDs
In the previous commit, the reply ID was added as a private state
variable to keep track of what post a person was replying to. In
some cases, the user may close the reply window and instead
make a new post entirely. This event flow causes the author view
to retain the reply ID, thus thinking it wanted to reply to that post
instead of making a new one.

This commit fixes this by resetting the reply ID when the view
disappears.