:children_crossing: Autofill respondents in replies
As stated in commit de71ac8, most Mastodon apps will autofill
the usernames of those interacting with the thread when the
user replies to a given post. The initial commit didn't include this
feature, which may make threads more confusing to follow.

This commit fixes this by adding an additional method when the
view appears to autofill in the respondents at the beginning of
the message.