:children_crossing: Make more general improvements
In the StatusDetailView, you can now see the statistical data for a
status and its replies, including the number of likes, boosts, and
replies. The user currently cannot contribute to the likes and boosts
number, but this may be changed in the future.

In the AuthorView, users can now add content warnings to indicate
that their status may be sensitive to their audience. They can also
provide a reason as to why it was marked sensitive. The AuthorView
also displays a more readable "Replying to <displayname>" in the
subtitle on macOS, and the character counter has been adjusted to
make sure the background stays in place.

The AuthorView also now prevents submission of statuses with more
than 500 characters in length by disabling the button to submit and
resetting the text to 600 characters if the user goes over 600
characters in the editor.