:sparkles: Add ability to like and reblog statuses
The latest changes to Chica allow developers to send requests to
like or boost a status. This functionality has been added to this
project with buttons that change their state based on likes and
boosts, running the appropriate action each time. The requests
return a new version of the status every time, making it easier to
refresh the view without needing to rewrite the code.

Additionally, some buttons in the UI now have appropriate help
tags associated with them, which appear as tooltips or info in
VoiceOver.