:sparkles: Add ability to toggle on passive activities
Passive Activities are described as activities that engage with a status
passively, such as liking and re-blogging. This commit introduces a
new settings view that will house settings for experimental features
such as the ability to turn off passive activities. Settings are stored in
the UserDefaults through the AppStorage wrapper.