:construction: Start work on messaging UI
The messaging UI will appear similar to other messaging apps
with bubbles on either side of the screen, depending on the
author of the message. Future commits will likely need to add
additional backend components such as rendering the
Conversation type in Mastodon, as well as adding contextual
data for a status.