:children_crossing: Make minor improvements to authoring
Because it is not possible to create a floating window that can be grabbed
and moved into a split view, the author buttons have been changed to
replace this functionality with a modal on iPadOS. More information can be
seen at the link below.

https://developer.apple.com/forums/thread/700820?answerId=705860022#705860022

The padding on the message texts in MessagePresentationView have been
adjusted to fit more content on the screen.

A new extension method Conversation.getAuthors(excluding:) has been
added to write the same author text seen in MessageListCellView; this is
used there, as well as the navigation title of the detail view.

AuthorView on iOS will be forced into a stack NavigationViewStyle instead
of a default assumed by the OS.

MessagePresentationView has been slightly refactored to reuse the avatar
code that was duplicated.

The MessageList view now includes a button to author a direct message
directly from that view, rather than heading to another page like "For You".
Consequently, on macOS, the visibility doesn't seem to be updated
correctly; however, the behavior acts as intended on iOS and iPadOS.