:construction: Add Send Feedback button component
The Send Feedback button will open the YouTrack bug reporter where
users can write feature requests and/or bug reports for any issues they
may find in the software. This button has been applied in both the iOS
and macOS apps, in spots that should be obvious to the user.