:sparkles: Update URL prefix from starlight:// to shout://
Previously, starlight:// was used to handle URL requests since the current
version of Chica responds to URLs with this specific type in mind.
However, a testing feature in hyperspacedev/chica#1 will make it possible
to set a new URL prefix on app launch. This commit uses this experimental
feature to set the URL prefix to shout:// instead of starlight://. If this feature
lands in the main Chica library, the Chica requirements will be updated
accordingly.