:children_crossing: Improve sign-in experience on iOS
In previous iterations, when the user wants to sign in to the app with Gopherdon,
they were redirected to Safari. However, this apparently violates the App Store
Guidelines, notably the "Design" section:

> We noticed that the user is taken to the default web browser to sign in or register
> for an account, which provides a poor user experience.

This commit resolves this by making a new SafariView available, ported over
from Give Me A Sniglet. When the user taps on the button to authenticate, the new
SafariView kicks in and opens the authorization URL. When the authorization
finishes, the view is automatically dismissed, and the user is directed to the For You
page. This is a step in the right direction, though it doesn't quite resolve SHU-1 yet.