:construction: Add mentions page
The official title for this page is still in progress, but the mentions
page shows all notifications where a user has replied to or mentions the
current user in a conversation.

Some of the layout code has been refactored into their own links, since
the GardensAppPage enum is pretty frequently used. It also helps ensure
consistency across various parts of the app.

Polls have also gotten an updated look to be more consistent with each
other.