:children_crossing: Update status nav list layout
The iconography for when a status has been bookmarked, favorited, and/or
reblogged by the user was in an awkward spot that didn't make the most
efficient use of the screen's space, especially on iPhone.

This has been remedied by changing it to be displayed under the status
itself with the Quick Glance view.