:recycle: Make improvements to composer text count
The composer's characters remaining count now more accurately reflects
the count as seen in the web version, only counting URLs at 23
characters instead of the full character amount. This is one of the
first steps for FGD-4.

The setting for health.interventions should now return true if the
setting isn't defined.