:bug: Fix 404 error on profile actions
The endpoints, although defined, were not pointing to their proper paths
in the Mastodon API. This has been fixed.

Additionally, the responses for these actions return a Relationship, and
not an Account. This also cuts down on calls to fetch the relationship,
so the code has been modified to utilize this behavior.