:recycle: Change Alice requests
Alice now utilizes a protocol to get a specified session and make a
request. This should make it incredibly easy to write more unit tests
for the app and Alice as a whole.

On top of this, Alice now introduces new methods to run GET, POST, PUT,
and DELETE requests more declaratively. Alice.request will still exist
for various reasons, but it is recommended to use those convenience
methods before resorting to Alice.request for readability.