:children_crossing: Add button to view attachments
The status detail view now has an option to show attachments. This item
is disabled when the status doesn't contain any media attachments, and
it is permanently attached to the toolbar.

The attachment viewer has been given some extra UX magic as well such as
showing alt text at the very bottom of the device, and a blurred version
of the media in the background.