:sparkles: Add poll editing to composer
The composer now has an option to add or edit polls. Poll editing
consists of setting an expiration date and the options that will be
displayed. When editing, the poll will be prepopulated with the rest of
the status.

Alice also properly supports query parameters with arrays as inputs,
thus making the POST and PUT calls for the status creation with polls
possible.

On top of this, a bug where the edit button would display on statuses
that the user has reblogged has been resolved.

Note that this functionality is limited to single-option polls, and that
multiple selections aren't supported yet.