:recycle: Update models
Account, Announcement, AnnouncementReaction, Application, Context,
Conversation, MediaAttachment, PreviewCard, and Status have been updated
to match more closely to the new API documentation. They are also now
value types, using a Box for recursive stored property as in the cases
of Account and Status. This change was made to better leverage Swift's
capabilities.

StatusView has also been updated as a result of this change to prevent
compilation errors with initializing the view.