:boom: Add AuthenticationModule
The authentication module handles the authentication flow presented by
TootSDK. Internally, it uses the ASWebAuthenticationSession to trigger a
login flow, rather than the prior method of running through an
SFSafariViewController.

It is important to note here that this will break the Alice package, as
the token generated for Toot SDK is not compatible with Alice.