:test_tube: Add bad URL mitigations
The URL object usually returns nil if the string it was generated from
contains invalid characters or is an empty string. It is unclear which
characters cause this issue, and I'd rather not implement telemetry.

Instead, this commit introduces a new handler to Alice.startOAuthFlow
that fires when the URL object is found to be nil, giving the user an
opportunity to file a report on Raceway to help diagnose FGD-22 more
clearly.