:recycle: Move Frugal Mode into package
Frugal Mode is now a feature module that resides in its own package that
controls the general aspects of the feature, including checking for Low
Power Mode. It includes its own Flow with FlowKit as well.

FlowKit also gains support for writing simpler test cases with a new
StatefulTestCase protocol that helps enforce specific checks with a flow
test.